RATIS-1214. Separate the default LogAppender implementation from LogAppenderBase.
